{topost} ACPI: resource: pass acpi dev to acpi_register_gsi()
Now when we register a gsi, it's impossible to make the
parent irq domain selectable, as the in the platform
device probed in acpi_platform.c, it will not pass
the acpi dev to register gsi, so introduce a adev
in struct res_proc_context to pass it as a context to
scan the interrupt resources, then finally pass to
acpi_register_gsi().

As former patch already interprete the interrupt-parent
property, so we can get the virq from diffrent domains.

Since we often pass NULL to acpi_register_gsi() and there
is no interrupt-parent property for devices connect to
GICD on ARM or io-apic ON X86, so it will use the default
irqdomain for those deivces and no functional changes to
those devices.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>